feat(database,service): Add user_id to result table and update related functions for user-specific data handling

+++ Go Framework/web-server/repo/delete_by_Id.go
@@ -2,9 +2,8 @@ package repo
 
 import "database/sql"
 
-
-func DeleteResultByID(db *sql.DB, id int) (int64, error) {
-	res, err := db.Exec(`DELETE FROM result WHERE id = $1`, id)
+func DeleteResultByID(db *sql.DB, id, userID int) (int64, error) {
+	res, err := db.Exec(`DELETE FROM result WHERE id = $1 AND user_id = $2`, id, userID)
 	if err != nil {
 		return 0, err
 	}
